requirements.txt
fastapi
uvicorn
sqlalchemy
pandas
pyodbc
python-dotenv
openai>=1.0.0
requests
